+++ defi/src/rwa/perps/platforms/adapters/extended.ts
@@ -1,5 +1,5 @@
-import type { PlatformAdapter, FundingEntry, ParsedPerpsMarket } from "./types";
-import { safeFloat } from "./types";
+import type { PlatformAdapter, FundingEntry, ParsedPerpsMarket } from "../types";
+import { safeFloat } from "../types";
 
 // Extended (formerly 10x Exchange) — StarkNet
 // Docs: https://docs.extended.exchange
