+++ defi/src/rwa/perps/platforms/adapters/gtrade.ts
@@ -1,5 +1,5 @@
-import type { PlatformAdapter, FundingEntry, ParsedPerpsMarket } from "./types";
-import { safeFloat, safeFetch } from "./types";
+import type { PlatformAdapter, FundingEntry, ParsedPerpsMarket } from "../types";
+import { safeFloat, safeFetch } from "../types";
 import { fetchPythPricesBySymbol } from "./pyth";
 import { applyOstiumFallbackPrices } from "./pyth";
