+++ defi/src/rwa/perps/platforms/adapters/pyth.ts
@@ -192,14 +192,14 @@ export async function fetchOstiumFallbackPrices(): Promise<Map<string, number>> 
  * defaults to extracting the base symbol from "venue:SYM-QUOTE" and appending "USD".
  */
 export async function applyOstiumFallbackPrices(
-  markets: import("./types").ParsedPerpsMarket[],
-  resolveKey?: (market: import("./types").ParsedPerpsMarket, ostiumPrices: Map<string, number>) => number,
+  markets: import("../types").ParsedPerpsMarket[],
+  resolveKey?: (market: import("../types").ParsedPerpsMarket, ostiumPrices: Map<string, number>) => number,
 ): Promise<void> {
   const missing = markets.filter((m) => m.markPx === 0);
   if (missing.length === 0) return;
 
   const ostiumPrices = await fetchOstiumFallbackPrices();
-  const defaultResolve = (m: import("./types").ParsedPerpsMarket) => {
+  const defaultResolve = (m: import("../types").ParsedPerpsMarket) => {
     const sym = m.contract.split(":")[1]?.split("-")[0] ?? "";
     return ostiumPrices.get(`${sym}USD`) ?? 0;
   };
